# openapi/backend.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories/create": {
      "post": {
        "operationId": "CategoryController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/goals/update-progress": {
      "put": {
        "operationId": "GoalController_updateProgress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "amount"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "amount": {
                    "type": "number",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Goals"
        ]
      }
    },
...
    "/goals/delete": {
      "delete": {
        "operationId": "GoalController_deleteGoal",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Goals"
        ]
      }
    },
...
    "/savings-boxes/stats": {
      "get": {
        "operationId": "SavingsBoxController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Savings Boxes"
        ]
      }
    },
...
    "/savings-boxes/create": {
      "post": {
        "operationId": "SavingsBoxController_createSavingsBox",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Savings Boxes"
        ]
      }
    },
...
    "/investments/update": {
      "put": {
        "operationId": "InvestmentController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvestmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentResponseEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Investments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Finance Control API",
    "description": "API de controle financeiro",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ft_access"
      }
    },
    "schemas": {
...
      "CreateCategoryCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Alimentação"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ],
            "example": "EXPENSE"
          },
          "color": {
            "type": "string",
            "example": "#FF5733"
          },
          "icon": {
            "type": "string",
            "example": "food"
          },
          "description": {
            "type": "string",
            "example": "Gastos com alimentação"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
...
      "InvestmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "renda_fixa",
              "acoes",
              "fundos",
              "fiis",
              "criptomoedas",
              "commodities",
              "internacional",
              "previdencia",
              "outros"
            ]
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "initial_amount": {
            "type": "number",
            "example": 1000
          },
          "current_amount": {
            "type": "number",
            "example": 1200
          },
          "target_amount": {
            "type": "number",
            "nullable": true,
            "example": 5000
          },
          "investment_date": {
            "type": "string",
            "format": "date"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time"
          },
          "is_active": {
            "type": "boolean"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "category",
          "initial_amount",
          "current_amount",
          "investment_date",
          "last_updated",
          "is_active",
          "created_at",
          "updated_at"
        ]
      },
...
      "InvestmentResponseEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/InvestmentResponseDto"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "UpdateInvestmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "renda_fixa",
              "acoes",
              "fundos",
              "fiis",
              "criptomoedas",
              "commodities",
              "internacional",
              "previdencia",
              "outros"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "renda_fixa",
              "acoes",
              "fundos",
              "fiis",
              "criptomoedas",
              "commodities",
              "internacional",
              "previdencia",
              "outros"
            ],
            "deprecated": true
          },
          "current_amount": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "target_amount": {
            "type": "number",
            "nullable": true,
            "minimum": 0.01
          },
          "investment_date": {
            "type": "string",
            "format": "date"
          },
          "is_active": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}